Create CNN model for road (road: .047 -> .265)
Created a fully convolutional model for identifying road pixels,
increasing the road F-score from .047 to .265 compared to the
previous heuristic model. Model predicts road for everything
that is not sky. Model was still improving rapidly at end of
2.5 minute training time, so further training is needed to
determine the strength of this model. Losses after each of
the five epochs were .46, .41, .37, .28, and .25. FPS dropped
from 17 to 8 compared to the heuristic model, below the
desired 10 FPS. Likely speed improvements include scaling,
cropping, and batching.