Display inferred probabilities as color bands
Improved the display of inferred probabilities using color
bands rather than grayscale. The improved display clarifies the
strengths and weaknesses of my road model. It recognizes almost
all of the road, except for some shadows and part of the yellow
lane line (common car color). Its biggest problem is inability
to distinguish road from sidewalk - both count as road. It does
not consider either a dark curb or a shadow cast on the road to
be road. Gray spots on the car's hood are also classified as
non-road. Overall, my impression is that this model is primarily
color-based.