Parallelize stages of image processing (FPS: 10 -> 20)
During submission, computation time is divided almost evenly
between three sub-tasks: pre-processing, applying predictive
models, and post-processing. One of those, applying predictive
models, is GPU-intensive. The other two are CPU-intensive. By
dividing these tasks into separate processes, the work can be
divided between the GPU and two CPUs for much higher FPS.

"done" messages are used to work around a bug in the
Pipe class, in which closing the pipe does not result in
the expected EOF for another process that reads from the
closed pipe.

FPS score throttled to around 10.3 to avoid hinting to other
students that this level of optimization is possible.