Add validation generator for car training
Added validation generator for car training to determine whether
the model was overfitting. Both the training and validation
generators start with the same base examples but apply random
augmentation independently. This approach showed no accuracy
difference between training and validation. That could either
indicate that the network is well regularized or that this
approach to building a validation set is insufficient.