Increase training batch size from 10 to 100
Increased training batch size from 10 to 100 to reduce random
variation in gradient and allow further reduction in the loss
with longer training times. Reached a loss of .0026 and
accuracy of .9947 in training. F-score dropped from .723 to
.699. Precision increased from .572 to .685. Recall dropped
from .774 to .703. Despite overall (small) drop in F-score,
I'm keeping this change because the training process should
be optimized to reduce the loss. If the loss doesn't
correlate well enough with this competition's scoring, I'll
deal with that as a separate issue.