Augment by copying vehicles between training images (worse result)
Augmented training data by copying vehicles from one training
example to another training example. The intention was to
increase both the representation and variety of vehicles in
the training data. (In the original training data, vehicles
represented 3% of total pixels.) This change resulted in a worse
car F-score (.723 -> .665) with an especially low precision of
37%. I am still confident in the idea, so am committing the
change with the intent of further tuning this augmentation to
be useful.