Add more training data from CARLA simulator (thanks chinkiat!)
Ong Chin-Kiat (chinkiat in SDC Slack) donated another 1000 training
examples from CARLA. I made minimal changes to the rest of the
project so that additional training examples could be included.
Car F-score dropped from .742 to .712. I am assuming that this
score change is a random variation due to retraining and that
my car model is not complex enough to take advantage of this
much training data.

Car F-score was .875 training, .779 validation, and .712 test.
Car precision was .601 training, .481 validation, and .560 test.
Car recall was .989 training, .923 validation, and .763 test.