Replace convolutions with inceptions (car: .742 -> .783, FPS: 23 -> 9)
Replaced each convolution in the car model with an inception
module of similar input/output depth. This increase in model
complexity increased the car F-score from .742 to .783 at the
cost of reducing speed from 23 to 8.9 FPS.

Car F-score was .912 training, .690 validation, and .993 test.
Car precision was .912 training, .704 validation, and .986 test.
Car recall was .783 training, .613 validation, and .841 test.