Profile the pre-processing step
Profiled the pre-processing step for signs that I could
optimize the pre-processing rather than trimming down
the models. Found that almost all pre-processing time
is spent on message passing. This means either that
message passing is actually slow, in which case I
need to explicitly encode arrays or switch channels,
or that sending a message triggers a context switch,
in which case pre-processing is a false lead.

Due to the uncertainty in this result, I'll probably
try trimming the road model next.

preprocessor :   113.9   (0.114 / frame) for 1000 frames
  pre:read :       3.2   (0.003 / frame) for 1000 frames
  pre:extract :    0.0   (0.000 / frame) for 1000 frames
  pre:cropscale : 13.6   (0.014 / frame) for 1000 frames
  pre:send :      97.1   (0.097 / frame) for 1000 frames
car infer :       55.6   (0.056 / frame) for 1000 frames
road infer :      54.2   (0.054 / frame) for 1000 frames
postprocess :      8.6   (0.009 / frame) for 1000 frames
encode array :     2.6   (0.003 / frame) for 1000 frames
encode png :      13.2   (0.013 / frame) for 1000 frames
encode json :      0.0   (0.000 / frame) for 1000 frames
total :          114.1   (0.114 / frame) for 1000 frames